Make URL input into a component
